Extracted job handler for convenience (related to https://github.com/camunda-community-hub/spring-zeebe/issues/132)
